Manejo de errores en los controllers

+++ recomendacionApp/recomendacionApp/src/main/java/cl/FilmFlux/recomendacionApp/controller/ResenaController.java
@@ -1,5 +1,6 @@
 package cl.FilmFlux.recomendacionApp.controller;
 
+import java.time.LocalDateTime;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -29,12 +30,29 @@ public class ResenaController {
 
     @GetMapping
     public ResponseEntity<List<Resena_DTO>> getResenas(){
-        return ResponseEntity.ok(resenaService.getResenas());
+        List<Resena_DTO> resenas = resenaService.getResenas();
+
+        if(resenas.isEmpty()){
+            System.out.println("[" + LocalDateTime.now() + "] No se encontraron reseñas");
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+        }
+
+        return ResponseEntity.ok(resenas);
     }
 
     @PostMapping
     public ResponseEntity<Resena> saveResena(@Valid @RequestBody Resena resena){
-        return ResponseEntity.status(HttpStatus.CREATED).body(resenaService.saveResena(resena));
+
+        if(resena == null){
+            System.out.println("[" + LocalDateTime.now() + "] Datos de reseña no proporcionados");
+            return ResponseEntity.badRequest().build();
+        }
+
+        Resena nuevaResena = resenaService.saveResena(resena);
+
+        return ResponseEntity
+                .status(HttpStatus.CREATED)
+                .body(nuevaResena);
     }
     
     @GetMapping("/{id}")
@@ -48,35 +66,57 @@ public class ResenaController {
     
     @PutMapping("/{id}")
     public ResponseEntity<Resena> actualizarResena(@PathVariable int id, @Valid @RequestBody Resena resena) {
-        resena.setIdResena(id);
-        Resena resenaActualizada = resenaService.updateResena(resena);
-        if (resenaActualizada == null) {
-            return ResponseEntity.notFound().build();
+        if(resena == null){
+            System.out.println("[" + LocalDateTime.now() + "] " + "Error al actualizar reseña: Datos de reseña no proporcionados");
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).build();
+        }else{
+            resena.setIdResena(id);
+            Resena resenaActualizada = resenaService.updateResena(resena);
+            if (resenaActualizada == null) {
+                return ResponseEntity.notFound().build();
+            }
+            return ResponseEntity.ok(resenaActualizada);
         }
-        return ResponseEntity.ok(resenaActualizada);
     }
     
     @DeleteMapping("/{id}")
     public ResponseEntity<Void> deleteResena(@PathVariable int id) {
-        resenaService.deleteResena(id);
-        return ResponseEntity.noContent().build();
+        if(resenaService.getResenaId(id) == null){
+             return ResponseEntity.notFound().build();
+        }else{
+            resenaService.deleteResena(id);
+            return ResponseEntity.noContent().build();
+        }        
     }
 
     // Metodos especiales
     @GetMapping("/resenasPorUsuario/{id}")
     public ResponseEntity<List<Resena_DTO>> getResenasByUsuario(@PathVariable int id){
-        return ResponseEntity.ok(resenaService.getResenasByUsuario(id));
+        List<Resena_DTO> resenas = resenaService.getResenasByUsuario(id);
+        if(resenas.isEmpty()){
+            System.out.println("[" + LocalDateTime.now() + "] No se encontraron reseñas para el usuario ID: " + id);
+            return ResponseEntity.notFound().build();
+        }
+        return ResponseEntity.ok(resenas);
     }
 
     @GetMapping("/resenasPorPelicula/{id}")
     public ResponseEntity<List<Resena_DTO>> getResenasByPelicula(@PathVariable int id){
-        return ResponseEntity.ok(resenaService.getResenasByPelicula(id));
+        List<Resena_DTO> resenas = resenaService.getResenasByPelicula(id);
+        if(resenas.isEmpty()){
+            System.out.println("[" + LocalDateTime.now() + "] No se encontraron reseñas para la película ID: " + id);
+            return ResponseEntity.notFound().build();
+        }
+        return ResponseEntity.ok(resenas);
     }
 
     @GetMapping("/resenasPorSerie/{id}")
     public ResponseEntity<List<Resena_DTO>> getResenasBySerie(@PathVariable int id){
-        return ResponseEntity.ok(resenaService.getResenasBySerie(id));
+        List<Resena_DTO> resenas = resenaService.getResenasBySerie(id);
+        if(resenas.isEmpty()){
+            System.out.println("[" + LocalDateTime.now() + "] No se encontraron reseñas para la serie ID: " + id);
+            return ResponseEntity.notFound().build();
+        }
+        return ResponseEntity.ok(resenas);
     }
-
-
 }
