+++ recomendacionApp/recomendacionApp/src/main/java/cl/FilmFlux/recomendacionApp/controller/SerieController.java
@@ -1,5 +1,6 @@
 package cl.FilmFlux.recomendacionApp.controller;
 
+import java.time.LocalDateTime;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -8,6 +9,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import cl.FilmFlux.recomendacionApp.model.Pelicula;
 import cl.FilmFlux.recomendacionApp.model.Serie;
 import cl.FilmFlux.recomendacionApp.service.SerieService;
 import jakarta.validation.Valid;
@@ -28,36 +30,50 @@ public class SerieController {
 
     @GetMapping
     public ResponseEntity<List<Serie>> getSeries(){
-        return ResponseEntity.ok(serieService.getSeries());
+        List<Serie> series = serieService.getSeries();
+
+        if(series.isEmpty()){
+            System.out.println("[" + LocalDateTime.now() + "] " + "No se encontraron series");
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+        }else{
+            return ResponseEntity.ok(serieService.getSeries());
+        }
     }
 
     @PostMapping
     public ResponseEntity<Serie> saveSerie(@Valid @RequestBody Serie serie){
-        return ResponseEntity.status(HttpStatus.CREATED).body(serieService.saveSerie(serie));
-    }
-    
-    @GetMapping("/{id}")
-    public ResponseEntity<Serie> getSerie(@PathVariable int id){
-        Serie serie = serieService.getSerieId(id);
         if(serie == null){
-            return ResponseEntity.notFound().build();
+            System.out.println("[" + LocalDateTime.now() + "] " + "Error al guardar serie: Datos de serie no proporcionados");
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).build();
+        }else{
+            return ResponseEntity.status(HttpStatus.CREATED).body(serieService.saveSerie(serie));
         }
-        return ResponseEntity.ok(serie);
     }
 
     @PutMapping("/{id}")
     public ResponseEntity<Serie> actualizarSerie(@PathVariable int id, @Valid @RequestBody Serie serie) {
-        serie.setIdSerie(id);
-        Serie serieActualizada = serieService.updateSerie(serie);
-        if (serieActualizada == null) {
-            return ResponseEntity.notFound().build();
+        if(serie == null){
+            System.out.println("[" + LocalDateTime.now() + "] " + "Error al actualizar serie: Datos de serie no proporcionados");
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).build();
+        }else{
+            serie.setIdSerie(id);
+            Serie serieActualizada = serieService.updateSerie(serie);
+            if (serieActualizada == null) {
+                return ResponseEntity.notFound().build();
+            }
+            return ResponseEntity.ok(serieActualizada);
         }
-        return ResponseEntity.ok(serieActualizada);
     }
     
-    @DeleteMapping("/{id}")
+   @DeleteMapping("/{id}")
     public ResponseEntity<Void> deleteSerie(@PathVariable int id) {
+
+        if(serieService.getSerieId(id) == null){
+             return ResponseEntity.notFound().build();
+        }
+
         serieService.deleteSerie(id);
+
         return ResponseEntity.noContent().build();
     }
